# shared/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BéninExpense AI - Plateforme de gestion des dépenses et conformité fiscale
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dgi/normalize/{expenseId}:
    post:
      operationId: normalizeExpense
      tags: [dgi]
      summary: Send expense to DGI for fiscal normalization (simulated)
      parameters:
        - name: expenseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Normalized with QR code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DgiResult"
  /dgi/status/{expenseId}:
    get:
      operationId: getDgiStatus
      tags: [dgi]
      summary: Get DGI normalization status for an expense
      parameters:
        - name: expenseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: DGI status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DgiResult"

  # Budgets
  /budgets:
    get:
      operationId: getBudgets
      tags: [budgets]
      summary: List budgets with current spending
      responses:
        "200":
          description: Budgets with spending data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BudgetWithSpending"
    post:
      operationId: createBudget
      tags: [budgets]
      summary: Create a budget
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BudgetInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetWithSpending"
  /budgets/{id}:
    patch:
      operationId: updateBudget
      tags: [budgets]
      summary: Update a budget
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                amount:
                  type: number
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetWithSpending"
    delete:
      operationId: deleteBudget
      tags: [budgets]
      summary: Delete a budget
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Employees
...
  /dgi/settings:
    get:
      operationId: getDgiSettings
      tags: [dgi]
      summary: Get DGI e-MECeF connection settings
      responses:
        "200":
          description: DGI settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DgiSettings"
    put:
      operationId: updateDgiSettings
      tags: [dgi]
      summary: Update DGI e-MECeF credentials
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DgiSettingsInput"
      responses:
        "200":
          description: Updated DGI settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DgiSettings"
components:
  schemas:
...
    BudgetInput:
      type: object
      required: [category, amount, period, year]
      properties:
        category:
          type: string
        amount:
          type: number
        period:
          type: string
          enum: [monthly, yearly]
        year:
          type: integer
        month:
          type: integer
    BudgetWithSpending:
      type: object
      required: [id, category, amount, period, year, spent, remaining, percentage, overBudget]
      properties:
        id:
          type: integer
        category:
          type: string
        amount:
          type: number
        period:
          type: string
          enum: [monthly, yearly]
        year:
          type: integer
        month:
          type: ["integer", "null"]
        spent:
          type: number
        remaining:
          type: number
        percentage:
          type: integer
        overBudget:
          type: boolean
        isCurrent:
          type: boolean

    # AI schemas
...
    DgiResult:
      type: object
      required: [expenseId, status]
      properties:
        expenseId:
          type: integer
        status:
          type: string
          enum: [pending, normalized, failed]
        qrCode:
          type: ["string", "null"]
        reference:
          type: ["string", "null"]
        normalizedAt:
          type: ["string", "null"]
          format: date-time

    # Employee schemas
...
    DgiSettings:
      type: object
      required: [configured, ifu, mode]
      properties:
        configured:
          type: boolean
          description: Whether real DGI credentials are set
        ifu:
          type: string
          description: Company IFU
        mode:
          type: string
          enum: [real, simulation]
          description: real if credentials configured, simulation otherwise
        defaultCurrency:
          type: string
          description: Default currency code (XOF, XAF, GHS, NGN…)
    DgiSettingsInput:
      type: object
      properties:
        dgiToken:
          type: string
          description: DGI e-MECeF API token (leave empty to clear)
        dgiSecret:
          type: string
          description: DGI e-MECeF API secret (leave empty to clear)
        defaultCurrency:
          type: string
          description: ISO 4217 currency code (XOF, XAF, GHS, NGN…)

    # ─── Boutiko schemas ───────────────────────────────────────────────────────
...